account/requirement.txt
django>=2.0,<3.0
django-cors-headers

django-admin-rangefilter
django-admin-list-filter-dropdown
django-admin-numeric-filter
django-monthfield
django-phonenumber-field[phonenumbers]
django-phonenumber-field[phonenumberslite]

django-pgcrypto

aiohttp
redis

django-redis

./apoweroftrance-django-utils-0.0.1.tar.gz

# RESTful API
djangorestframework

# Swagger 2.0
drf-yasg

# Database
psycopg2-binary

# JWT
pyjwt

# AES
pycrypto

fleep
